Began testing with testPropagateMagField. Altered ChawlaSharma to divide input momentum by mass (mass is constant for pure magnetic field), and then multiply by mass on output. Altered testPropagateMagField to initialize fEquation to be Mag_UsualEqRhs_IntegrateByTime instead of G4Mag_UsualEqRhs. Changed G4ChordFinder to initialize with Mag_UsualEqRhs_IntegrateByTime instead of G4Mag_UsualEqRhs. Copied G4ChargeState from newer version of Geant4 (there was a bug with get and set methods for PDGSpin, which were not present in the older version of Geant4). CMakeLists.txt file is different in test/ directory. It no longer includes the directory test/testSteppers (you can just change it back in order to compile modifiedtestH.cc).

+++ test/testSteppers/testH.cc
@@ -28,7 +28,7 @@ int main(int argc, char *args[]){
 	if(argc>1)
 		stepper_no = atoi(args[1]);
 	if(argc > 2)
-		step_len = stof(args[2]);
+		step_len = atof(args[2]);
 
 
 	//1. Create a field :
